+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Unless otherwise specified, any version comparison below is the comparison of th
 ### Added
 
 - (API) Added new permission `Permissions.COMMAND_VIEW_OTHER_OUTPUTS`, player now must have that permission to be able to view other's command outputs.
+- (API) Added event `PlayerStartFishEvent` that will be called when a player tries to throw a fishing rod.
 - Added `max-decompressed-bytes` network setting to `server-settings.yml` to control the maximum decompressed packet size (default 50 MB).
 - Added `NettyPipelineInitEvent` for packet level middleware support. The `AllayPlayer` instance is attached to the channel attribute.
 - Added pitch and yaw parameters to the teleport command.